Remove duplicate assignment ducks
