+++ packages/opensrp-plans/package.json
@@ -55,6 +55,7 @@
     "@opensrp/plan-form-core": "^0.0.3",
     "@opensrp/react-utils": "^0.0.7",
     "@opensrp/server-service": "^0.0.17",
+    "@opensrp/team-assignment": "^0.0.2",
     "@opensrp/team-management": "^0.0.5",
     "react-i18next": "^11.8.10"
   },
