+++ packages/opensrp-plans/src/components/AssignmentTable/index.tsx
@@ -11,11 +11,11 @@ import {
 } from '@opensrp/team-management';
 import {
   assignmentsReducer,
-  assignmentReducerName,
+  assignmentsReducerName,
   Assignment,
   fetchAssignments,
   getAssignmentsArrayByPlanId,
-} from '../../ducks/assignments';
+} from '@opensrp/team-assignment';
 import {
   JurisdictionsReducer,
   jurisdictionReducerName,
@@ -40,7 +40,7 @@ import {
   makePlanDefinitionsArraySelector,
 } from '../../ducks/planDefinitions';
 
-reducerRegistry.register(assignmentReducerName, assignmentsReducer);
+reducerRegistry.register(assignmentsReducerName, assignmentsReducer);
 reducerRegistry.register(orgReducerName, organizationsReducer);
 reducerRegistry.register(jurisdictionReducerName, JurisdictionsReducer);
