+++ packages/opensrp-plans/src/components/AssignmentTable/tests/index.test.tsx
@@ -19,7 +19,7 @@ import {
 import { act } from 'react-dom/test-utils';
 import { jurisdictions } from './fixtures';
 import { Jurisdiction, removeJurisdictions } from '../../../ducks/jurisdictions';
-import { removeAssignmentsAction } from '../../../ducks/assignments';
+import { removeAssignmentsAction } from '@opensrp/team-assignment';
 import { Organization, removeOrganizationsAction } from '@opensrp/team-management';
 import lang from '../../../lang';
 import { Dictionary } from '@onaio/utils';
