+++ packages/opensrp-plans/src/components/AssignmentTable/tests/utils.test.tsx
@@ -1,4 +1,4 @@
-import { processRawAssignments } from '../../../ducks/assignments/utils';
+import { processRawAssignments } from '@opensrp/team-assignment';
 import { compressAssignments, getDataSource, mergeIdsWithNames } from '../utils';
 import {
   expectedMergedIds1,
