+++ packages/opensrp-plans/src/components/AssignmentTable/utils.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { ColumnsType, ColumnType } from 'antd/lib/table/interface';
 import { TableColumnsNamespace } from '../../constants';
 import { SelectOption } from '../AssignmentModal';
-import { Assignment, fetchAssignments } from '../../ducks/assignments';
+import { Assignment, fetchAssignments } from '@opensrp/team-assignment';
 import { Dictionary, keyBy } from 'lodash';
 import { Organization } from '@opensrp/team-management';
 import { Jurisdiction } from '../../ducks/jurisdictions';
