+++ packages/opensrp-plans/src/index.tsx
@@ -2,7 +2,6 @@ import '@opensrp/plan-form/dist/index.css';
 export * from './constants';
 export * from './components/PlanInfo';
 export * from './ducks/planDefinitions';
-export * from './ducks/assignments';
 export * from './ducks/jurisdictions';
 export * from './containers/ListView';
 export * from './containers/NewPlan';
