+++ packages/team-assignment/src/components/ListView/index.tsx
@@ -22,11 +22,12 @@ import {
 import { PlanDefinition } from '@opensrp/plan-form-core';
 import { Helmet } from 'react-helmet';
 import { useDispatch, useSelector } from 'react-redux';
-import reducer, {
+import {
+  assignmentsReducer,
   Assignment,
   fetchAssignments,
-  getAssignments,
-  reducerName as assignmentReducerName,
+  assignmentsReducerName,
+  getAssignmentsArrayByPlanId,
 } from '../../ducks/assignments';
 import {
   ASSIGNMENTS_ENDPOINT,
@@ -41,7 +42,9 @@ const { fetchAllHierarchies, getAllHierarchiesArray } = locationHierachyDucks;
 
 reducerRegistry.register(orgReducerName, organizationsReducer);
 reducerRegistry.register(locationHierachyDucks.reducerName, locationHierachyDucks.reducer);
-reducerRegistry.register(assignmentReducerName, reducer);
+reducerRegistry.register(assignmentsReducerName, assignmentsReducer);
+
+const assignmentsSelector = getAssignmentsArrayByPlanId();
 
 export interface TableData {
   id: string;
@@ -77,7 +80,9 @@ const TeamAssignmentView = (props: TeamAssignmentViewProps) => {
   const Treedata = useSelector(
     (state) => (getAllHierarchiesArray(state) as unknown) as ParsedHierarchyNode[]
   );
-  const assignmentsList: Assignment[] = useSelector((state) => getAssignments(state));
+  const assignmentsList: Assignment[] = useSelector((state) =>
+    assignmentsSelector(state, { planId: defaultPlanId })
+  );
   const allOrganizations: Organization[] = useSelector((state) => getOrganizationsArray(state));
   const dispatch = useDispatch();
   const [loading, setLoading] = useState<boolean>(true);
@@ -180,11 +185,11 @@ const TeamAssignmentView = (props: TeamAssignmentViewProps) => {
 
   const tableData = dataSource.map((datum: ParsedHierarchyNode, i: number) => {
     const jurisdictionAssignments = assignmentsList.filter(
-      (assignment) => assignment.jurisdictionId === datum.id
+      (assignment) => assignment.jurisdiction === datum.id
     );
     const jurisdictionOrgs = allOrganizations.filter((org) => {
       const jurisdictionOrgIds = jurisdictionAssignments.map(
-        (assignment) => assignment.organizationId
+        (assignment) => assignment.organization
       );
       return jurisdictionOrgIds.includes(org.identifier);
     });
