+++ packages/team-assignment/src/index.tsx
@@ -1,1 +1,3 @@
 export * from './components/ListView';
+export * from './ducks/assignments';
+export * from './ducks/assignments/utils';
